Fix sending server memory state at compilation failure
PR #11628 introduced a bug where it was assumed that message type
is the same as compilation failures status code, as described in #11711.

This commit fixes it by adding a second parameter to
`ServerStream::writeError`,
and using it to send the server state.

+++ runtime/compiler/runtime/CompileService.cpp
@@ -46,5 +46,5 @@ void J9CompileDispatcher::compile(JITServer::ServerStream *stream)
          }
       } // end critical section
    // If we reached this point there was a memory allocation failure
-   stream->writeError(compilationLowPhysicalMemory, JITServer::ServerMemoryState::VERY_LOW);
+   stream->writeError(compilationLowPhysicalMemory, (uint64_t) JITServer::ServerMemoryState::VERY_LOW);
    }
